Remove the use caddy
